Reorganise EFL review components
Namespace these components for ease of discovery and to make it obvious
they're all related to the same feature.
