+++ spec/services/candidate_interface/choose_efl_review_component_spec.rb
@@ -8,7 +8,7 @@ RSpec.describe CandidateInterface::ChooseEflReviewComponent do
       it 'returns an instance of IeltsReviewComponent' do
         component = described_class.call(english_proficiency)
 
-        expect(component).to be_instance_of(IeltsReviewComponent)
+        expect(component).to be_instance_of(CandidateInterface::EnglishForeignLanguage::IeltsReviewComponent)
         expect(component.ielts_qualification).to eq english_proficiency.efl_qualification
       end
     end
@@ -19,8 +19,19 @@ RSpec.describe CandidateInterface::ChooseEflReviewComponent do
       it 'returns an instance of ToeflReviewComponent' do
         component = described_class.call(english_proficiency)
 
-        expect(component).to be_instance_of(ToeflReviewComponent)
+        expect(component).to be_instance_of(CandidateInterface::EnglishForeignLanguage::ToeflReviewComponent)
         expect(component.toefl_qualification).to eq english_proficiency.efl_qualification
+      end
+    end
+
+    context 'when english_proficiency has an Other EFL qualification' do
+      let(:english_proficiency) { build(:english_proficiency, :with_other_qualification) }
+
+      it 'returns an instance of OtherEflQualificationReviewComponent' do
+        component = described_class.call(english_proficiency)
+
+        expect(component).to be_instance_of(CandidateInterface::EnglishForeignLanguage::OtherEflQualificationReviewComponent)
+        expect(component.other_qualification).to eq english_proficiency.efl_qualification
       end
     end
 
@@ -30,7 +41,7 @@ RSpec.describe CandidateInterface::ChooseEflReviewComponent do
       it 'returns an instance of NoEflQualificationReviewComponent' do
         component = described_class.call(english_proficiency)
 
-        expect(component).to be_instance_of(NoEflQualificationReviewComponent)
+        expect(component).to be_instance_of(CandidateInterface::EnglishForeignLanguage::NoEflQualificationReviewComponent)
         expect(component.english_proficiency).to eq english_proficiency
       end
     end
